ENV: moving flight compiler, main webserver database updater to utils and serbserver respectively
