+++ src/backend/server/database_updater.go
@@ -1,4 +1,4 @@
-package flightdb
+package fffwebserver 
 
 
 import (
