+++ utils/data/process/compile/flights/get-all-valid-destinations.go
